requirements-dev.txt
-r requirements.txt
django-debug-toolbar
Fabric3
flake8
ipython
mypy
coverage
django_coverage_plugin
coveralls
python-coveralls
factory_boy
tox